requirements.txt
# python==3.11.11
# ipykernel==6.29.5
# ipywidgets==8.1.5

pandas==2.2.3
tqdm==4.67.1
torch==2.6.0
transformers==4.48.3
torchvision==0.21.0
decord==0.6.0
einops==0.8.1
sentencepiece==0.2.0
wheel==0.45.1 # req to install flash-attention
flash_attn==2.7.4 # also add this while installing "--no-build-isolation"
accelerate-1.3.0-py # for qwen-2.5-14B
openai==1.68.2
google_genai==1.7.0
timeout-decorator==0.5.0

scikit_learn==1.6.1